+++ tests/unit/utils.py
@@ -7,12 +7,18 @@ import yaml
 
 import circuit_build
 
-TEST_DIR = Path(__file__).resolve().parent
-TEST_DATA_DIR = TEST_DIR / "proj66-tiny"
-TEST_DATA_DIR_SYNTH = TEST_DIR / "proj66-tiny-synth"
-TEST_DATA_DIR_NGV = TEST_DIR / "ngv"
+
+UNIT_TESTS_DIR = Path(__file__).resolve().parent
+TEST_DATA_DIR = UNIT_TESTS_DIR / "proj66-tiny"
+TEST_DATA_DIR_SYNTH = UNIT_TESTS_DIR / "proj66-tiny-synth"
+
 SNAKEFILE = pkg_resources.resource_filename(circuit_build.__name__, "snakemake/Snakefile")
-SNAKEMAKE_ARGS = ["--bioname", str(TEST_DATA_DIR), "-u", str(TEST_DATA_DIR / "cluster.yaml")]
+SNAKEMAKE_ARGS = [
+    "--bioname",
+    str(TEST_DATA_DIR),
+    "-u",
+    str(TEST_DATA_DIR / "cluster.yaml"),
+]
 
 
 @contextmanager